dev: gcdb: display: check for msm chip serial number
When an APPS bootloader image is flashed on device and then
subsequently the msm is replaced, the dfps structure stored in
splash partition is not valid. The DSI PLL trim codes need to be
generated again for the new msm part. Save the chip serial number
which is unique for each msm part along with the dfps struct in
splash partition. If the current chip serial number does not match
the chip serial number in the splash partition, recalculate the
DSI PLL trim codes and store it back to splash partition.

Change-Id: I8476d60454c52ee9adbbb6386c6e6bb91c666911
Signed-off-by: Yasir Malik <ymalik@codeaurora.org>