target: msm8996: Reprogram the correct QSEED phase values
During switching from early camera to animated splash and vice-versa,
camera layer is made transparent and opaque respectively. When switching
back to opaque layer QSEED phase values were programmed wrongly because
of which there was a blue flicker. So to fix this issue the phase values which
were calculated during the pipe configuration are written back to the registers
while making the camera layer opaque.

Change-Id: I357f40bc2368d04ad1fc03a2fb0d293827cc44ef
Acked-by: Poojashree Masthi <c_pmasth@qti.qualcomm.com>
Signed-off-by: Rahul Sharma <rahsha@codeaurora.org>