target: msm8996: Reprogram the Pixel Extension registers.
Blue line was observed at the bottom of the early camera frame due to improper
programming of the Pixel Extension registers. These registers are configured
in mdp_scalar_config and are reprogrammed in mdss_source_pipe_config function.
The issue is solved by adding static function which is invoked by both the
above mentioned functions.

Change-Id: I95f4004a06475c766c05604a4ed349e4989b0420
Acked-by: Poojashree Masthi <c_pmasth@qti.qualcomm.com>
Signed-off-by: Rahul Sharma <rahsha@codeaurora.org>