bugfix: do not crash when pressing volume keys
At 05fff3ba8ce839451b0185ec6fe855739f88bf8d QC tried (either on
purpose or by mistake) to have the bootloader no longer react to
KEY_VOL_UP and KEY_VOL_DOWN on a couple of sittuations (notably, in
the verity screens).

QC's approach is buggy - it simply doesn't define a function for those
keypresses, but an attempt to execute it still happens, and with it
the bootloader fails and crashes.

I see no reason why not let the menu be surfaced with those buttons,
so I'm reverting to that behavior, but if we want the buttons to not
do anything we can, too - we'll just need to do it properly.

Change-Id: Ib4d07de2bb7394bcd4355ab08b887cffd73b6fbb